refactor(home page features): rename folders to change case (Step 2 of 2)
