+++ src/features/HomePage/index.js
@@ -1,3 +1,3 @@
-export * from "./components/hero-section2";
-export * from "./components/search-section2";
-export * from "./components/top-recipes2";
+export * from "./components/hero-section";
+export * from "./components/search-section";
+export * from "./components/top-recipes";
